[Oztechan/CCC#1826] Move DateUtil into :client:core:shared
