+++ client/src/commonTest/kotlin/com/oztechan/ccc/client/mapper/ExchangeRateTest.kt
@@ -1,8 +1,8 @@
 package com.oztechan.ccc.client.mapper
 
+import com.oztechan.ccc.client.core.shared.util.nowAsInstant
+import com.oztechan.ccc.client.core.shared.util.toDateString
 import com.oztechan.ccc.client.helper.BaseTest
-import com.oztechan.ccc.client.util.nowAsInstant
-import com.oztechan.ccc.client.util.toDateString
 import com.oztechan.ccc.common.core.model.Conversion
 import com.oztechan.ccc.common.core.model.ExchangeRate
 import kotlin.test.Test
