+++ client/src/commonTest/kotlin/com/oztechan/ccc/client/util/PremiumUtilTest.kt
@@ -1,5 +1,7 @@
 package com.oztechan.ccc.client.util
 
+import com.oztechan.ccc.client.core.shared.util.nowAsLong
+import com.oztechan.ccc.client.core.shared.util.toInstant
 import com.oztechan.ccc.client.model.PremiumType
 import kotlinx.datetime.DateTimePeriod
 import kotlinx.datetime.TimeZone
