+++ client/src/commonTest/kotlin/com/oztechan/ccc/client/viewmodel/WidgetViewModelTest.kt
@@ -1,12 +1,12 @@
 package com.oztechan.ccc.client.viewmodel
 
+import com.oztechan.ccc.client.core.shared.util.isItOver
+import com.oztechan.ccc.client.core.shared.util.nowAsLong
 import com.oztechan.ccc.client.datasource.currency.CurrencyDataSource
 import com.oztechan.ccc.client.helper.BaseViewModelTest
 import com.oztechan.ccc.client.service.backend.BackendApiService
 import com.oztechan.ccc.client.storage.app.AppStorage
 import com.oztechan.ccc.client.storage.calculator.CalculatorStorage
-import com.oztechan.ccc.client.util.isItOver
-import com.oztechan.ccc.client.util.nowAsLong
 import com.oztechan.ccc.client.viewmodel.widget.WidgetViewModel
 import com.oztechan.ccc.common.core.infrastructure.constants.DAY
 import com.oztechan.ccc.common.core.model.Conversion
